+++ src/js/shared/Prop/MockProps.mjs
@@ -2,9 +2,9 @@ import Mock from "../Transport/MockTransport.mjs"
 
 const mocks = {}
 
-function mock(module, method, args, def) {
+function mock(module, method, params, def) {
   const fullMethod = `${module}.${method}`
-  if ((args == null) || args.length === 0 || (Object.values(args[0]).length === 0)) {
+  if ((params == null) || (Object.values(params).length === 0)) {
     // get
     const rv = mocks[fullMethod] && (mocks[fullMethod].value != null) ? mocks[fullMethod].value : def
     return rv
@@ -17,11 +17,11 @@ function mock(module, method, args, def) {
       }
     }
     mocks[fullMethod] = mockMethod
-    mockMethod.value = args[0].value
+    mockMethod.value = params.value
     Mock.event(module, method + 'Changed', {
-      value: args[0].value
+      value: params[0].value
     })
-    return {}
+    return null
   }
 }