+++ src/template/js/methods/default.js
@@ -1,7 +1,4 @@
 
 function ${method.name}(${method.params}) {
-
-  const transforms = ${method.transforms}
-
-  return Transport.send('${info.title}', '${method.name}', { ${method.params} }, transforms)
+  return Transport.send('${info.title}', '${method.name}', { ${method.params} }, ${transport.options})
 }